feat(api): add typed path strategy
