+++ src/conformly/__init__.py
@@ -3,7 +3,7 @@ from ._internal.parser.adapters.bootstrap import register_builtin_adapters
 register_builtin_adapters()
 
 # ruff: noqa: E402
-from ._internal.api import case, cases
+from ._internal.api import case, cases, path
 from ._internal.constraints import (
     GreaterOrEqual,
     GreaterThan,
@@ -19,6 +19,7 @@ from ._internal.constraints import (
     UniqueItems,
 )
 from ._internal.fields import Email, IPv4, IPv6, IPvAny
+from ._internal.types import ViolationType as V
 from .exceptions import ConformlyError
 
 __all__ = [
@@ -39,6 +40,8 @@ __all__ = [
     "OneOf",
     "Pattern",
     "UniqueItems",
+    "V",
     "case",
     "cases",
+    "path",
 ]
