+++ src/conformly/_internal/api/_errors.py
@@ -1,0 +1,7 @@
+from typing import Any
+
+from conformly.exceptions import GenerationError
+
+
+def api_error(message: str, *, code: str, **context: Any) -> GenerationError:
+    return GenerationError(message, context={"code": code, **context})
